Added Admin-owner UI(fetch,verify,delete,reject)-owner

+++ residentia-backend/src/main/java/com/example/demo/controller/DashboardController.java
@@ -23,8 +23,8 @@ public class DashboardController {
         this.bookingService = bookingService;
     }
 
-    private Long getLoggedOwnerId() {
-        return 1L;
+    private Integer getLoggedOwnerId() {
+        return (int) 1L;
     }
 
     @GetMapping
